+++ src/flowrep/nodes/__init__.py
@@ -11,7 +11,7 @@ from flowrep.nodes import (
     while_recipe,
     workflow_recipe,
 )
-from flowrep.nodes.union import RecipeDiscrimination, Recipes
+from flowrep.nodes.union_types import RecipeDiscrimination, Recipes
 
 # Subtlety: Anywhere we use `typing.TYPE_CHECKING` to avoid a real import _and_ use
 # the guarded object as a pydantic field annotator, we are going to need to make sure
