+++ src/flowrep/nodes/if_recipe.py
@@ -8,7 +8,7 @@ from flowrep import base_models, edge_models, subgraph_validation
 from flowrep.nodes import helper_models
 
 if TYPE_CHECKING:
-    from flowrep.nodes.union import Recipes
+    from flowrep.nodes.union_types import Recipes
 
 
 class IfRecipe(base_models.NodeRecipe):
