+++ src/flowrep/parsers/parser_protocol.py
@@ -6,7 +6,7 @@ from typing import Protocol, runtime_checkable
 from pyiron_snippets import versions
 
 from flowrep import edge_models
-from flowrep.nodes import union, workflow_recipe
+from flowrep.nodes import union_types, workflow_recipe
 from flowrep.parsers import object_scope, symbol_scope
 
 
@@ -17,7 +17,7 @@ class BodyWalker(Protocol):
     scope: object_scope.ScopeProxy
     symbol_map: symbol_scope.SymbolScope
     info_factory: versions.VersionInfoFactory
-    nodes: union.Recipes
+    nodes: union_types.Recipes
 
     @property
     def inputs(self) -> list[str]: ...
