+++ tests/unit/nodes/test_union_types.py
@@ -11,7 +11,7 @@ from flowrep.nodes import (
     for_recipe,
     if_recipe,
     try_recipe,
-    union,
+    union_types,
     while_recipe,
     workflow_recipe,
 )
@@ -251,7 +251,7 @@ class TestDiscriminatedUnionRoundtrip(unittest.TestCase):
 
     def test_discriminator_resolves_correct_type(self):
         """Each node type is correctly identified via 'type' discriminator."""
-        adapter = pydantic.TypeAdapter(union.RecipeDiscrimination)
+        adapter = pydantic.TypeAdapter(union_types.RecipeDiscrimination)
         for type_enum, data, expected_class in self.test_cases:
             with self.subTest(type=type_enum.value):
                 node = adapter.validate_python(data)
@@ -260,7 +260,7 @@ class TestDiscriminatedUnionRoundtrip(unittest.TestCase):
 
     def test_roundtrip_through_union(self):
         """Serialize then deserialize through the union type."""
-        adapter = pydantic.TypeAdapter(union.RecipeDiscrimination)
+        adapter = pydantic.TypeAdapter(union_types.RecipeDiscrimination)
         for type_enum, data, expected_class in self.test_cases:
             with self.subTest(type=type_enum.value):
                 node = adapter.validate_python(data)
@@ -271,7 +271,7 @@ class TestDiscriminatedUnionRoundtrip(unittest.TestCase):
 
     def test_json_schema_includes_all_types(self):
         """Union schema should reference all node types."""
-        adapter = pydantic.TypeAdapter(union.RecipeDiscrimination)
+        adapter = pydantic.TypeAdapter(union_types.RecipeDiscrimination)
         schema = adapter.json_schema()
         # Discriminated unions use anyOf or oneOf
         self.assertTrue(
@@ -286,17 +286,19 @@ class TestDiscriminatedUnionRoundtrip(unittest.TestCase):
             "inputs": ["a", "b"],
             "outputs": ["c"],
         }
-        node = pydantic.TypeAdapter(union.RecipeDiscrimination).validate_python(data)
+        node = pydantic.TypeAdapter(union_types.RecipeDiscrimination).validate_python(
+            data
+        )
         self.assertIsInstance(node, atomic_recipe.AtomicRecipe)
         self.assertEqual(node.reference.inputs_with_defaults, ["b"])
 
         # Full roundtrip
-        dumped = pydantic.TypeAdapter(union.RecipeDiscrimination).dump_python(
+        dumped = pydantic.TypeAdapter(union_types.RecipeDiscrimination).dump_python(
             node, mode="json"
         )
-        restored = pydantic.TypeAdapter(union.RecipeDiscrimination).validate_python(
-            dumped
-        )
+        restored = pydantic.TypeAdapter(
+            union_types.RecipeDiscrimination
+        ).validate_python(dumped)
         self.assertEqual(restored.reference.inputs_with_defaults, ["b"])
 
     def test_workflow_roundtrip_preserves_defaults(self):
@@ -310,7 +312,9 @@ class TestDiscriminatedUnionRoundtrip(unittest.TestCase):
             "edges": {},
             "output_edges": {},
         }
-        node = pydantic.TypeAdapter(union.RecipeDiscrimination).validate_python(data)
+        node = pydantic.TypeAdapter(union_types.RecipeDiscrimination).validate_python(
+            data
+        )
         self.assertIsInstance(node, workflow_recipe.WorkflowRecipe)
         self.assertEqual(node.reference.inputs_with_defaults, ["x"])
 
@@ -319,7 +323,7 @@ class TestDiscriminatorValidation(unittest.TestCase):
     """Tests for discriminator error handling."""
 
     def test_unknown_type_rejected(self):
-        adapter = pydantic.TypeAdapter(union.RecipeDiscrimination)
+        adapter = pydantic.TypeAdapter(union_types.RecipeDiscrimination)
         with self.assertRaises(pydantic.ValidationError) as ctx:
             adapter.validate_python(
                 {
@@ -332,7 +336,7 @@ class TestDiscriminatorValidation(unittest.TestCase):
         self.assertIn("type", exc_str.lower())
 
     def test_missing_type_rejected(self):
-        adapter = pydantic.TypeAdapter(union.RecipeDiscrimination)
+        adapter = pydantic.TypeAdapter(union_types.RecipeDiscrimination)
         with self.assertRaises(pydantic.ValidationError):
             adapter.validate_python(
                 {
@@ -344,7 +348,7 @@ class TestDiscriminatorValidation(unittest.TestCase):
 
     def test_type_mismatch_with_fields_rejected(self):
         """Type says 'atomic' but fields are for workflow."""
-        adapter = pydantic.TypeAdapter(union.RecipeDiscrimination)
+        adapter = pydantic.TypeAdapter(union_types.RecipeDiscrimination)
         with self.assertRaises(pydantic.ValidationError):
             adapter.validate_python(
                 {
@@ -378,7 +382,7 @@ class TestNodesTypeAlias(unittest.TestCase):
                 "outputs": [],
             },
         }
-        adapter = pydantic.TypeAdapter(union.Recipes)
+        adapter = pydantic.TypeAdapter(union_types.Recipes)
         nodes = adapter.validate_python(nodes_data)
         self.assertEqual(len(nodes), 2)
         self.assertIsInstance(nodes["step1"], atomic_recipe.AtomicRecipe)
@@ -394,7 +398,7 @@ class TestNodesTypeAlias(unittest.TestCase):
                 "outputs": [],
             },
         }
-        adapter = pydantic.TypeAdapter(union.Recipes)
+        adapter = pydantic.TypeAdapter(union_types.Recipes)
         with self.assertRaises(pydantic.ValidationError) as ctx:
             adapter.validate_python(nodes_data)
         self.assertIn("for", str(ctx.exception))
@@ -411,7 +415,7 @@ class TestNodesTypeAlias(unittest.TestCase):
                         "outputs": [],
                     },
                 }
-                adapter = pydantic.TypeAdapter(union.Recipes)
+                adapter = pydantic.TypeAdapter(union_types.Recipes)
                 with self.assertRaises(pydantic.ValidationError):
                     adapter.validate_python(nodes_data)
 
@@ -442,13 +446,13 @@ class TestNodesTypeAlias(unittest.TestCase):
             },
             "workflow_node": inner_workflow,
         }
-        adapter = pydantic.TypeAdapter(union.Recipes)
+        adapter = pydantic.TypeAdapter(union_types.Recipes)
         nodes = adapter.validate_python(nodes_data)
         self.assertIsInstance(nodes["atomic_node"], atomic_recipe.AtomicRecipe)
         self.assertIsInstance(nodes["workflow_node"], workflow_recipe.WorkflowRecipe)
 
     def test_empty_nodes_dict(self):
-        adapter = pydantic.TypeAdapter(union.Recipes)
+        adapter = pydantic.TypeAdapter(union_types.Recipes)
         nodes = adapter.validate_python({})
         self.assertEqual(nodes, {})
 
@@ -487,7 +491,7 @@ class TestNestedUnionResolution(unittest.TestCase):
             "edges": {},
             "output_edges": {"results": "for_node.out"},
         }
-        adapter = pydantic.TypeAdapter(union.RecipeDiscrimination)
+        adapter = pydantic.TypeAdapter(union_types.RecipeDiscrimination)
         wf = adapter.validate_python(data)
         self.assertIsInstance(wf, workflow_recipe.WorkflowRecipe)
         self.assertIsInstance(wf.nodes["for_node"], for_recipe.ForEachRecipe)
@@ -521,7 +525,7 @@ class TestNestedUnionResolution(unittest.TestCase):
             "edges": {},
             "output_edges": {"out": "middle.b"},
         }
-        adapter = pydantic.TypeAdapter(union.RecipeDiscrimination)
+        adapter = pydantic.TypeAdapter(union_types.RecipeDiscrimination)
         wf = adapter.validate_python(outer)
         self.assertIsInstance(wf.nodes["middle"], workflow_recipe.WorkflowRecipe)
         self.assertIsInstance(
